Serverless/function-source/requirements.txt
functions-framework==3.*
requests==2.24.0
scikit-learn==1.2.2
google-cloud-storage==1.25.0
numpy==1.19.2